# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi Docs with swagger",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "ouldhenniaabdelkader@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/blog": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get all posts",
                "description": "Returns a list of all posts.",
                "operationId": "getPosts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "thread"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "thread"
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "example": null
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": null
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-04-08T20:39:27.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/create": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Post Blog",
                "description": "Returns blog Information with 201 status response code.",
                "operationId": "Post Blog",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Title of your BLog"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Body of your Blog"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "thread"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "thread"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blog/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get a specific your Blog by ID",
                "description": "Get blog by Id",
                "operationId": "getYourModelById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of your Blog",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Get Blog By Id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "thread"
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "thread"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete/{id}": {
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Delete Blog by ID",
                "description": "Get blog by Id",
                "operationId": "Get Model By Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of your Blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Delete the blog with Id Sussecfly"
                    }
                }
            }
        },
        "api/edit/{id}": {
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update a your Blog",
                "description": "Update a your Blog by its ID",
                "operationId": "update the blog by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of your Blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Abdelkader"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Ouldn Hennia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "thread"
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "thread"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}